refactor: layout index and infos
