+++ tailwind.config.ts
@@ -18,6 +18,9 @@ export default {
       "500": "#799880",
       "600": "#6B7F6E",
       "700": "#5C665C",
+      "800": "#243628",
+      "900": "#1D241E",
+      "950": "#1E211F",
       red: "#FF5050",
       redHover: "#E65E5E",
       gray: "#6B7280",
